Fix memory issue in index import command
PHP was having issues managing memory in the import method. This change puts all
the import batch logic into it's own method, which allows the php
garbage collector to do it's job better. The memory still climbs as we
reindex (a problem to fix later), but this drastically reduces the
amount of memory used each loop.